restyled front page as welcome only
