+++ src/App.css
@@ -23,3 +23,37 @@
 .center{
   text-align: center;
 }
+
+/* home */
+html, body {
+    margin: 0;
+    height: 100%;
+}
+
+.image{
+  position: relative;
+  width: 100%;
+}
+
+.overlay {
+  position: absolute;
+  top: 100px;
+  left: 0;
+  width: 100%;
+}
+
+.massive {
+  font-size: 5vw;
+}
+
+.center {
+  text-align: center;
+}
+
+
+.about {
+  position: absolute;
+  top: 100px;
+  left: 100;
+  width: 30%;
+}
