+++ src/features/activities/ActivitiesData.js
@@ -1,0 +1,41 @@
+  const programming = {
+    id: 1,
+    title: "Computer Programming Education",
+    subtitle: "One Step at a Time",
+    description: "With gender equity always in mind, we strive to find new strategies for dealing with the challenge of advancing global economic development. Computer literacy is something that we take very seriously, and our team is working each and every day to make a positive impact. Contact us to learn more about our commitment to this cause.",
+    buttonText: "Help us Succeed",
+    buttonLink: 'https://www.paypal.com/cgi-bin/webscr?cmd=_donations&business=info@codigoecuador.com&item_name=Imagina%20C%C3%B3digo%20Ecuador&currency_code=USD',
+    image: "https://theadelantemovement.com/wp-content/uploads/2016/05/22543243101_46a15d19e9_b-800x540.jpg"
+  }
+
+  const internship = {
+    id: 2,
+    title: "Internship Placement",
+    subtitle: "Experiential Learning",
+    description: "At Código Ecuador, we are dedicated to connecting Ecuadorian high school students to rewarding hands-on experiences. Through cooperation and global community empowerment, we believe we can facilitate access to valuable learning opportunities. We are always striving to make a difference, and invite you to learn more and lend your support.",
+    buttonText: "Become a Partner",
+    buttonLink: "mailto:info@codigoecuador.com",
+    image: 'https://www.wola.org/wp-content/uploads/2016/04/Diego-and-Interns-e1477074226989.jpg'
+
+  }
+
+const entrepreneurship = {
+  id: 3,
+  title: "Encourage Entrepreneurship",
+  subtitle: "One Step at a Time",
+  description: "With our organization’s mission always in mind, we strive to find new strategies to reach our goal. To preserve Ecuador's beautiful culture of entrepreneurship, Código Ecuador works to tear down walls, pave roads to resources, and encourage women to build opportunity. Contact us to learn more about how we make this happen.",
+  buttonText: "Get in Touch",
+  buttonLink:"/contact",
+  image: "https://www.imf.org/external/pubs/ft/fandd/2019/03/images/032019/lagarde-600.jpg"
+}
+const entrepreneurship2 = {
+  id: 3,
+  title: "Encourage Entrepreneurship",
+  subtitle: "One Step at a Time",
+  description: "With our organization’s mission always in mind, we strive to find new strategies to reach our goal. To preserve Ecuador's beautiful culture of entrepreneurship, Código Ecuador works to tear down walls, pave roads to resources, and encourage women to build opportunity. Contact us to learn more about how we make this happen.",
+  buttonText: "Get in Touch",
+  buttonLink:"/contact",
+  image: "https://www.imf.org/external/pubs/ft/fandd/2019/03/images/032019/lagarde-600.jpg"
+}
+
+export const activities = [programming, internship, entrepreneurship,entrepreneurship2]
